Remap ctrl-g to show the _full_ path to the current file
I usually want to see the full path to the current file as opposed to
just the basename. This could be annoying if I'm in a really deeply
nested directory structure, but that's not usual, and I don't think the
cognative load of skimming to the end of the path is too bad.