polybar: add janky indicator of which pulseaudio sink is active in the bar
Ideally I'd be able to just tell polybar to reload my pulseaudio modules
specifically, but individual module reload isn't currently supported. For now
I'll hard-restart the bar with IPC. It makes the bar flash off and then back
on in an ugly manner, but I think that's the best I can do for now.