polybar: change play/pause click listeners to trigger generic play/pause action
I think I originally had it only triggering for Firefox specifically because
I was using the Spotify web player, and didn't want to have the "play" action
trigger a YouTube video or Twitch stream. Having used it for a while though,
I think it's more useful to just have it perform the generic play/pause action.